requirements.txt
numpy
onnxruntime
opencv-python-headless
pycuda
tensorrt
torch
django
django-jazzmin